+++ src/pages/Product.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export const Product = () => {
+  return (
+    <div>
+      <h1>Helloo product</h1>
+    </div>
+  )
+};
