+++ src/pages/Settings.jsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+export const Settings = () => {
+  return (
+    <div>
+      <h1>Hello Settings</h1>
+    </div>
+  )
+};
+
+
