+++ src/pages/Users.jsx
@@ -1,0 +1,12 @@
+import React from 'react'
+
+export const Users = () => {
+  return (
+    <div>
+      <h1>Hello Users</h1>
+    </div>
+  )
+};
+
+
+
